Added way to specify unsupported query parameters and provide a warning
- Also moved tests around

</br>
