+++ tests/server/middleware/test_query_param.py
@@ -57,3 +57,80 @@ def test_empty_parameters(both_clients):
         query_part
     )
     assert expected_result == parsed_set_of_queries
+
+
+def test_wrong_query_param(check_error_response):
+    request = "/structures?_exmpl_filter=nelements=2"
+    check_error_response(
+        request,
+        expected_status=400,
+        expected_title="Bad Request",
+        expected_detail="The query parameter(s) '['_exmpl_filter']' are not recognised by this endpoint.",
+    )
+
+    request = "/structures?filer=nelements=2"
+    check_error_response(
+        request,
+        expected_status=400,
+        expected_title="Bad Request",
+        expected_detail="The query parameter(s) '['filer']' are not recognised by this endpoint.",
+    )
+
+    request = "/structures/mpf_3819?filter=nelements=2"
+    check_error_response(
+        request,
+        expected_status=400,
+        expected_title="Bad Request",
+        expected_detail="The query parameter(s) '['filter']' are not recognised by this endpoint.",
+    )
+
+
+def test_handling_prefixed_query_param(check_response):
+    request = "/structures?_odbx_filter=nelements=2&filter=elements LENGTH >= 9"
+    expected_ids = ["mpf_3819"]
+    check_response(request, expected_ids)
+
+    request = (
+        "/structures?_unknown_filter=elements HAS 'Si'&filter=elements LENGTH >= 9"
+    )
+    expected_ids = ["mpf_3819"]
+    expected_warnings = [
+        {
+            "title": "UnknownProviderQueryParameter",
+            "detail": "The query parameter(s) '['_unknown_filter']' are unrecognised and have been ignored.",
+        }
+    ]
+    check_response(
+        request, expected_ids=expected_ids, expected_warnings=expected_warnings
+    )
+
+
+def test_unsupported_optimade_query_param(check_response):
+
+    request = "/structures?filter=elements LENGTH >= 9&page_number=1"
+    expected_ids = ["mpf_3819"]
+    expected_warnings = [
+        {
+            "title": "QueryParamNotUsed",
+            "detail": "The query parameter(s) '['page_number']' are not supported by this server and have been ignored.",
+        }
+    ]
+    check_response(
+        request, expected_ids=expected_ids, expected_warnings=expected_warnings
+    )
+
+    request = "/structures?filter=elements LENGTH >= 9&page_number=1&_unknown_filter=elements HAS 'Si'"
+    expected_ids = ["mpf_3819"]
+    expected_warnings = [
+        {
+            "title": "UnknownProviderQueryParameter",
+            "detail": "The query parameter(s) '['_unknown_filter']' are unrecognised and have been ignored.",
+        },
+        {
+            "title": "QueryParamNotUsed",
+            "detail": "The query parameter(s) '['page_number']' are not supported by this server and have been ignored.",
+        },
+    ]
+    check_response(
+        request, expected_ids=expected_ids, expected_warnings=expected_warnings
+    )
